backend/auth/requirements.txt
Django>=4.0,<5.0
djangorestframework
requests
gunicorn
psycopg2-binary  # For PostgreSQL